requirements.txt
crewai
pydantic
python-dotenv
litellm